Update upgrade-packages.sh
- General updates
- Remove dev2 and dev3 python environments

+++ upgrade-packages.sh
@@ -1,13 +1,13 @@
 #!/usr/bin/env bash
-# Upgrade system (homebrew) and Python packages
+# Upgrade system packages
 
 
-cd "$(dirname "$0")"
+cd "$(dirname "$0")" || exit
 
 
-# Process Script Arguments
+# Script arguments
 all=true
-for i in ${@}; do
+for i in "${@}"; do
     case ${i} in
         brew)
         brew=true
@@ -29,16 +29,6 @@ for i in ${@}; do
         all=
         ;;
 
-        dev2)
-        dev2=true
-        all=
-        ;;
-
-        dev3)
-        dev3=true
-        all=
-        ;;
-
         conda)
         conda=true
         all=
@@ -53,95 +43,71 @@ if [[ ${brew} ]] || [[ ${all} ]]; then
     brew doctor
     brew upgrade
     brew cleanup
-
-    # printf "\nDowngrading Fish to v2.7.1; while v3.0 is broken"
-    # # To be fixed in Fish v3.1?
-    # # https://github.com/fish-shell/fish-shell/issues/5456
-    # # https://github.com/pypa/pipenv/issues/3414
-    # brew unlink fish
-    # brew install https://raw.githubusercontent.com/Homebrew/homebrew-core/2827b020c3366ea93566a344167ba62388c16c7d/Formula/fish.rb
-    # brew link fish
-
-
-    printf "\nUpdating pyenv shims"
-    pyenv rehash
 fi
 
 
 if [[ ${poetry} ]] || [[ ${all} ]]; then
     printf "\n==> Intalling/upgrading Python Poetry\n"
-    if [[ $(poetry --version 2>/dev/null) ]]; then
+    if command -v poetry 1>/dev/null 2>&1; then
         printf "Updating poetry..."
         poetry self update
     else
         printf "Installing poetry..."
         curl -sSL https://raw.githubusercontent.com/python-poetry/poetry/master/get-poetry.py | python -
     fi
-    source $HOME/.poetry/env
+
+    source "$HOME/.poetry/env"
+
     printf "Enabling poetry tab completion\n"
     # Bash (Homebrew)
-    poetry completions bash > $(brew --prefix)/etc/bash_completion.d/poetry.bash-completion
-
-    # Fish
-    poetry completions fish > ~/.config/fish/completions/poetry.fish
-
+    poetry completions bash > "$(brew --prefix)/etc/bash_completion.d/poetry.bash-completion"
     # Oh-My-Zsh
-    mkdir -pv $HOME/.oh-my-zsh/custom/plugins/poetry
-    poetry completions zsh > $HOME/.oh-my-zsh/custom/plugins/poetry/_poetry
+    mkdir -pv "$HOME/.oh-my-zsh/custom/plugins/poetry"
+    poetry completions zsh > "$HOME/.oh-my-zsh/custom/plugins/poetry/_poetry"
 fi
 
 
 if [[ ${sys2} ]] || [[ ${all} ]]; then
-    printf "\n==> Performing a clean install of the Python2 System Packages\n"
+    printf "\n==> Performing a clean install of the python2 system packages\n"
 
-    sys2_packages=`mktemp -t sys2_packages`
+    sys2_packages=$(mktemp -t sys2_packages)
 
     printf "\nUninstalling existing Python2 (pip2) System Packages"
-    pip2 freeze > ${sys2_packages}
-    pip2 uninstall -y -r ${sys2_packages}
+    pip2 freeze > "${sys2_packages}"
+    pip2 uninstall -y -r "${sys2_packages}"
 
-    sort -uo python/sys2-requirements.txt python/sys2-requirements.txt
-
-    printf "\nInstalling Python2 (pip2) System Packages from python/sys2-requirements.txt"
+    printf "\nInstalling pip2 setup tools"
     pip2 install --upgrade pip setuptools wheel
-    pip2 install --upgrade -r python/sys2-requirements.txt
 
-    rm ${sys2_packages}
+    if [[ -f "python/sys2-requirements.txt" ]]; then
+        printf "\nInstalling python2 (pip2) system packages from python/sys2-requirements.txt"
+        sort -uo python/sys2-requirements.txt python/sys2-requirements.txt
+        pip2 install --upgrade -r python/sys2-requirements.txt
+    fi
+
+    rm "${sys2_packages}"
 fi
 
 
 if [[ ${sys3} ]] || [[ ${all} ]]; then
-    printf "\n==> Upgrading Python3 System Packages\n"
+    printf "\n==> Performing a clean install of the python3 system packages\n"
 
-    sys3_packages=`mktemp -t sys3_packages`
+    sys3_packages=$(mktemp -t sys3_packages)
 
     printf "\nUninstalling existing Python3 (pip3) System Packages"
-    pip3 freeze > ${sys3_packages}
-    pip3 uninstall -y -r ${sys3_packages}
+    pip3 freeze > "${sys3_packages}"
+    pip3 uninstall -y -r "${sys3_packages}"
 
-    sort -uo python/sys3-requirements.txt python/sys3-requirements.txt
-
-    printf "\nInstalling Python3 (pip3) System Packages from python/sys3-requirements.txt"
+    printf "\nInstalling pip3 setup tools"
     pip3 install --upgrade pip setuptools wheel
-    pip3 install --upgrade -r python/sys3-requirements.txt
 
-    rm ${sys3_packages}
-fi
+    if [[ -f "python/sys3-requirements.txt" ]]; then
+        printf "\nInstalling python3 system packages from python/sys3-requirements.txt"
+        sort -uo python/sys3-requirements.txt python/sys3-requirements.txt
+        pip3 install --upgrade -r python/sys3-requirements.txt
+    fi
 
-
-if [[ ${dev2} ]] || [[ ${all} ]]; then
-    printf "\n==> Performing a clean build of the Python dev2 Virtual Environment from python/dev2-requirements.txt\n"
-    sort -uo python/dev2-requirements.txt python/dev2-requirements.txt
-    pew rm dev2
-    pew new -d -p python2 -r python/dev2-requirements.txt dev2
-fi
-
-
-if [[ ${dev3} ]] || [[ ${all} ]]; then
-    printf "\n==> Performing a clean build of the Python dev3 Virtual Environment from python/dev3-requirements.txt\n"
-    sort -o python/dev3-requirements.txt python/dev3-requirements.txt
-    pew rm dev3
-    pew new -d -p python3 -r python/dev3-requirements.txt dev3
+    rm "${sys3_packages}"
 fi
